Dataflash : Added wingtip logging (untested)
Fixed compiling errors.  All seems to work now.

Added wintip updates to the 50 Hz schedualed loop

Impoved the example script

Added stuff so it prints as a union

Got the sample script working with two wingtip boards.  Hangs really badly when two are required to be used.

Just working on the wingtip library

Added a checksum to the wingtip boards